requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==4.2.27
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==26.0
PyJWT==2.11.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
